+++ R/shinyGizmo-package.R
@@ -19,6 +19,17 @@ NULL
       if (length(.) == 0) NA else unlist(.)
     })
   }, force = TRUE)
+  shiny::registerInputHandler("shinyGizmo.valuebutton", function(x, shinysession, name) {
+    if (!is.null(x$type)) {
+      handler <- `%:::%`("shiny", "inputHandlers")$get(x$type)
+      if (!is.null(handler)) {
+        return(handler(x$value))
+      } else {
+        return(x$value)
+      }
+    }
+    return(x)
+  }, force = TRUE)
 }
 
 `%:::%` <- function (pkg, name) {