+++ NEWS.md
@@ -1,5 +1,7 @@
 # shinyGizmo 0.3
 
+* Add `try_binding` argument to `valueButton`. When `TRUE` and `selector` points to Shiny Binding and `attribute == "value"`
+it tries to convert sourced input value using registered `inputHandler`.
 * Add `jsCallOncePerFlush` function. When used prevents running `conditionalJS` callback during a single flush cycle ([3668](https://github.com/rstudio/shiny/issues/3668)).
 * Add `commonInput` and `commonInputs` functions that allow to gather input from multiple controllers into one.
 * Add `mergeCalls` function that allows to use more than one `jsCalls` for `conditionalJS`.